+++ src/MovieTrailers/MovieTrailers.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { fetchMovieTrailers } from '../apiCalls';
 import { Carousel } from 'react-responsive-carousel';
 import ReactPlayer from 'react-player';
-import "react-responsive-carousel/lib/styles/carousel.min.css"
+import 'react-responsive-carousel/lib/styles/carousel.min.css';
 import './MovieTrailers.scss';
 
 class MovieTrailers extends Component {
