+++ src/apiCalls/apiCalls.test.js
@@ -1,0 +1,2 @@
+import { fetchMovies, fetchSingleMovie, fetchMovieTrailers, postLoginCredentials, fetchUserRatings, postUserRating, deleteUserRating } from './apiCalls';
+import { user, singleMovie, trailers, movies }
